+++ src/ledgers/mod.rs
@@ -82,3 +82,115 @@ pub mod prelude {
     pub use super::single_ledger_request::*;
 }
 
+#[cfg(test)]
+pub mod tests {
+    use super::prelude::*;
+    use crate::horizon_client::HorizonClient;
+    use base64::{engine::general_purpose, Engine};
+
+    #[tokio::test]
+    async fn test_get_all_ledgers() {
+        let hash = "f96c4021adc1ae496c662f4f97143e499a9548f541c64bb2401a1b1701de5150";
+        let prev_hash = "63d98f536ee68d1b27b5b89f23af5311b7569a24faf1403ad0b52b633b07be99";
+
+        // Initialize horizon client
+        let horizon_client =
+            HorizonClient::new("https://horizon-testnet.stellar.org".to_string()).unwrap();
+
+        // construct request
+        let all_ledgers_request = LedgersRequest::new().set_limit(2).unwrap();
+
+        let all_ledgers_response = horizon_client.get_all_ledgers(&all_ledgers_request).await;
+        assert!(all_ledgers_response.is_ok());
+
+        let binding = all_ledgers_response.unwrap();
+        let all_ledgers_response = &binding.embedded().records()[0];
+
+        assert_eq!(all_ledgers_response.hash(), hash);
+        assert_eq!(all_ledgers_response.prev_hash(), prev_hash);
+        assert_eq!(all_ledgers_response.sequence(), &2);
+        assert_eq!(all_ledgers_response.successful_transaction_count(), &0);
+        assert_eq!(all_ledgers_response.paging_token(), "8589934592");
+    }
+
+    #[tokio::test]
+    async fn test_get_single_ledger() {
+        let id = "f96c4021adc1ae496c662f4f97143e499a9548f541c64bb2401a1b1701de5150";
+        let hash = "f96c4021adc1ae496c662f4f97143e499a9548f541c64bb2401a1b1701de5150";
+        let prev_hash = "63d98f536ee68d1b27b5b89f23af5311b7569a24faf1403ad0b52b633b07be99";
+        let closed_at = "2024-02-06T17:32:26Z";
+        let closed_at_timepoint = 1707240746;
+
+        // Initialize horizon client
+        let horizon_client =
+            HorizonClient::new("https://horizon-testnet.stellar.org".to_string()).unwrap();
+
+        // construct request
+        let single_ledger_request = SingleLedgerRequest::new().set_sequence(2).unwrap();
+
+        let single_ledger_response = horizon_client
+            .get_single_ledger(&single_ledger_request)
+            .await;
+
+        assert!(single_ledger_response.is_ok());
+        let single_ledger_response = single_ledger_response.unwrap();
+        assert_eq!(single_ledger_response.id(), id);
+        assert_eq!(single_ledger_response.paging_token(), "8589934592");
+        assert_eq!(single_ledger_response.hash(), hash);
+        assert_eq!(single_ledger_response.prev_hash(), prev_hash);
+        assert_eq!(single_ledger_response.sequence(), &2);
+        assert_eq!(single_ledger_response.successful_transaction_count(), &0);
+        assert_eq!(single_ledger_response.failed_transaction_count(), &0);
+        assert_eq!(single_ledger_response.operation_count(), &0);
+        assert_eq!(single_ledger_response.tx_set_operation_count(), &0);
+        assert_eq!(single_ledger_response.closed_at(), closed_at);
+        assert_eq!(single_ledger_response.total_coins(), "100000000000.0000000");
+        assert_eq!(single_ledger_response.fee_pool(), "0.0000000");
+        assert_eq!(single_ledger_response.base_fee_in_stroops(), &100);
+        assert_eq!(single_ledger_response.base_reserve_in_stroops(), &100000000);
+        assert_eq!(single_ledger_response.max_tx_set_size(), &100);
+        assert_eq!(single_ledger_response.protocol_version(), &0);
+
+        let decoded_xdr_header = single_ledger_response.decoded_header_xdr().unwrap();
+
+        assert_eq!(
+            decoded_xdr_header.bucket_list_hash.to_string(),
+            "735227ed398461291237687b08446aa2c9b096e0c98a462dadda569f05dd2484"
+        );
+
+        assert_eq!(decoded_xdr_header.ledger_seq, 2);
+        assert_eq!(decoded_xdr_header.total_coins, 1000000000000000000);
+        assert_eq!(decoded_xdr_header.fee_pool, 0);
+        assert_eq!(decoded_xdr_header.inflation_seq, 0);
+        assert_eq!(decoded_xdr_header.id_pool, 0);
+        assert_eq!(decoded_xdr_header.base_fee, 100);
+        assert_eq!(decoded_xdr_header.base_reserve, 100000000);
+        assert_eq!(decoded_xdr_header.max_tx_set_size, 100);
+
+        let tx_set_hash = decoded_xdr_header.scp_value.tx_set_hash.to_string();
+        let tx_set_hash_bytes = hex::decode(tx_set_hash.clone()).expect("Failed to decode hex");
+        let tx_set_hash_base64 = general_purpose::STANDARD.encode(tx_set_hash_bytes.clone());
+
+        assert_eq!(
+            tx_set_hash_base64,
+            "uZRHr9UdXKbTKiclfOjy72YZFJUkJPVcKT5htvorm1Q="
+        );
+
+        assert_eq!(
+            decoded_xdr_header.scp_value.close_time,
+            stellar_xdr::curr::TimePoint(closed_at_timepoint)
+        );
+
+        assert_eq!(
+            decoded_xdr_header.ext,
+            stellar_xdr::curr::LedgerHeaderExt::V0
+        );
+        for decoded in decoded_xdr_header.skip_list {
+            assert_eq!(
+                decoded.to_string(),
+                "0000000000000000000000000000000000000000000000000000000000000000"
+            );
+        }
+    }
+}
+
